Continued doing so denoizing.
We have done different things with Pamina, and I really have no clue what is best for our type of data. I would say we can go and discuss this with Milo next week. I am not sure but he might be on holidays so we can always send him an email.